requirements.txt
discord.py>=2,<3
discord-ui>=5,<6
python-dotenv>=0.21,<0.22